WIP: TemplateGUI tests
